stackeval: Record all applied actions
Instead of synthesizing a compound action from a sequence of two, here
we store a slice of zero or more actions applied to a resource instance
object. This can then be used directly in the calling context to
calculate a count of applied actions for the component.

Note that we have no test coverage for the change to the component
instance caller of the hook, which will follow in a later commit.