lang: sensitive value in dynamic for_each errors
When expanding a dynamic block with a for_each expression, the for_each
value is not permitted to contain sensitive values. Prior to this
change, the error returned to the user did not mention the sensitivity
of the value, instead saying only that the collection was not iterable.

Using new functionality introduced in hcl/dynblock, we can now check
whether the for_each value is sensitive and return a more helpful error.