stackeval: Another module source test case
If multiple submodules of a given component share the same relative
source, they should result in distint source addresses in the
diagnostics. This commit introduces an example with the following
structure:

- component.remote_invalid_grandchildren
  - module.first_child
    - module.child (source = "./child")
  - module.second_child
    - module.child (source = "./child")

Both of these use the same invalid module as the rest of the examples in
this test, so we should (and do) see equivalent diagnostics differing
only in filename.