Stacks: Update stack change hooks to send provider address in RPC messages
This updates the `ReportResourceInstanceStatus` and
`ReportResourceInstancePlanned` callbacks to extract the provider address they
received in their arguments and include it in the RPC messages they emit.

The corresponding RPC message formats were recently updated to include a field
for provider address.

The planned callback already could have accessed the provider address prior to
this PR, but the status callback needed the recent refactor of the
`terraform.Hook` API.