website: First draft of Terraform v1.8 upgrade guide
We may revise this further during the Terraform v1.8 prerelease period, if
feedback suggests that more information is needed or that this information
is not clear enough. The main goal here is just to get out of the confusing
situation where the v1.8 branch contains a v1.7 upgrade guide, so that
we can link to the upgrade guide from the changelog.

The full upgrade guide page gives us room for more details, allowing the
changelog to serve just as a concise list of changes.