configs: Don't treat module_variable_optional_attrs as special
Due to some specific circumstances around how this experiment reached
stabilization we made a planned-temporary exception to make it valid to
specify even in non-alpha builds of Terraform, just to be able to return
a more precise error message.

The comment added at the time suggested that we remove this "a release or
two after v1.3", and v1.8 is the actively-developed release at the time
I'm writing this so we've kept this around far longer than we originally
felt we needed to. It's time for it to go.

This particular experiment will now get the same treatment as all others
when used in a non-alpha release: Terraform will return the general error
about experiments only being available in alpha releases.