experiments: removed_provisioners
This new experiment keyword represents being able to declare provisioners
in "removed" blocks, so that it's possible to retain a destroy-time
provisioner even when the resource it belonged to is no longer declared
in the configuration.

However, that's not yet implemented as of this commit. The implementation
associated with this keyword will follow in subsequent commits.