go.mod: Eliminate github.com/apparentlymart/go-dump
I originally wrote this library as a stopgap measure because
davecgh/go-spew produced a hard-to-read representation of cty.Value due to
exposing the raw internals, whereas my "go-dump" instead just pretty-prints
whatever fmt.Sprintf("%#v", ...) would return for a given value.

However, in the meantime I wrote ctydebug as a more cty-specialized
library for this purpose, and we switched to using go-cmp as our primary
library for deep-comparing data structures, and so our few remaining uses
of apparentlymart/go-dump were vestigial in tests that haven't been
updated recently enough to have been updated to our new standards.

This commit therefore swaps out all use of that library with combinations
of go-cmp and ctydebug, matching our current standards for how to do these
sorts of things in tests. This removes our dependency on "go-dump"
altogether.

At least one of our dependencies is still using go-dump as part of its own
test suite, and so the Go toolchain still includes its declared version
in our go.sum file. However, this library is no longer linked in to
Terraform either in real builds or in test binaries, and so it gets dropped
entirely from our go.mod file.