backend: new "backendbase" helper package
This is intended as a lighter-weight replacement for the Backend type in
the legacy/helper/schema package, which indirectly pulls in our local copy
of the entire legacy Terraform plugin SDK.

This depends on the slightly-lower-level configschema package, which means
that it's not quite as feature-rich as what it's replacing, but backends
have relatively modest needs and so we can compensate with just some extra
logic in their Configure functions, which will be considerably simpler than
the huge bunch of legacy code we're currently depending on to get the same
effects. In particular, configschema has no concept of "default values"
for arguments, and so we'll need to deal with those in the Configure
methods using imperative code, instead of declaratively.

Nothing is using this yet, but in future commits we'll gradually convert
all of the remote state backends to use this instead of the legacy system,
and then remove the legacy system.