backendbase: BoolValue and MustBoolValue
With legacy helper/schema it's common to just type-assert whatever
ResourceData.Get returns and trust that it'll always be valid. We can't
achieve something quite that lazy with cty, but we can at least factor
out the work of turning a value we already know is a bool into a bool
value.

We also have a "must" variant for this one, because it's relatively easy
for a caller to ensure that it controls all of the inputs enough to
ensure that it should never fail.