backendbase: "SDK-like" helpers
A lot of the remote state backends' current Configure functions
unfortunately rely quite heavily on the "do the best you can" API of the
legacy SDK, which makes questionable shortcuts like treating null as
equivalent to the Go zero value of a type, but in return makes it
convenient to assign into other data structures that happen to make
similar assumptions. And of course that means that the backends' other
data structures were built to make those similar assumptions!

This new set of utilities therefore aims to provide a bare-minimum subset
of "SDK-like" functionality to ease the initial migration away from the
legacy SDK. It isn't exactly equivalent an in particular focuses only on
the small subset of legacy SDK functionality that most existing backends
rely on, but this should at least make the initial adaptations a little
more mechanical so that we can continue to defer the full modernization
of these backends for some undetermined future time without having to
continue retaining the huge legacy SDK snapshot in this codebase.