backendbase: Basic support for default values
Some existing backends make very heavy use of environment variable-based
default values and fallback values, which are available in the legacy SDK
but not in Terraform's own schema model.

To ease the migration away from the legacy SDK for those backends,
backendbase.Base now supports a lightweight mechanism for declaring such
defaults, and applies them automatically as part of PrepareConfig. This
gets us pretty close to the assumptions that the SDK allowed callers to
make, but with considerably less code. It's intentionally considerably
less flexible than the legacy SDK was, because in practice most backends
use only a tiny subset of the SDK functionality anyway.