backend/kubernetes: Remove legacy helper/schema dependency
As with most of our remote state backends, this one was depending on just
a tiny slice of the (enormous and now-poorly-understood) legacy SDK.

In an effort to eliminate the legacy SDK snapshot from this codebase, this
replaces it with functionality from our new "backendbase" package, which
aims to provide just a narrow set of utilities to minimize the churn
caused by removing the legacy SDK and thus reduce the risk of this change.

This is currently using the "SDK-like" utilities, which emulate some of the
questionable-but-convenient assumptions the legacy SDK makes, such as
the idea that empty string and null are equivalent. Hopefully in future
we can wean this backend even further off of these older assumptions, but
the priority for now is to eliminate the legacy SDK without significantly
disturbing the shape of the existing working code.