tfdiags: Support errors.Join and stdlib wrapped errors
HashiCorp needed support for wrapping multiple errors into a single error
and for wrapping one error inside another long before that was in the
standard library, and so built its own libraries for those needs.

However, the standard library now has its own answers for both and so the
HashiCorp libraries are redundant. We've already eliminated all use of
both elsewhere in Terraform (except for some legacy packages that predate
tfdiags anyway), so we no longer need to support them in package tfdiags.