dependencies: Remove all non-legacy uses of hashicorp/go-multierror
This functionality is now provided by the "errors" package in the standard
library, so the HashiCorp-specific library is obsolete.

This also removed our one direct use of github.com/pkg/errors, which we
were using in a way that was totally redundant with the stdlib package
errors. However, it lives on as an indirect dependency through the Consul
backend's dependency on the Consul SDK, which uses that library.