dependencies: No more golang.org/x/exp
This module was a temporary home for what are now the standard library
"slices" and "maps" packages. We're building Terraform with a sufficiently
new version of Go that we can use the standard library equivalents, so we
no longer need this dependency.

It lives on as an indirect dependency for now because
hashicorp/aws-sdk-go-base is also using it, but hopefully it'll vanish
altogether in future if that library is updated to also use stdlib instead.