copy: DeepCopyValue utility
Over the years we've gradually removed all use of various third-party
reflection-based libraries for deep-copying and projecting data structures
from main code, because they proved to be a maintenance nightmare and thus
better replaced by direct code that is less fun to write but far clearer
to read.

However, we've still got a few examples of _test_ code hanging on to these
helpers because the tradeoffs for unit tests tend to be different. That's
valid, but it's annoying to have to include several large and very general
dependencies just to support some tests whose needs are well-known and
relatively simple.

Therefore this new helper function copy.DeepCopyValue aims to be a more
straightforward replacement for mitchellh/copystructure (which also
indirectly subsumes mitchellh/reflectwalk) that eschews all of the
customizability that those libraries offered and instead focuses narrowly
on what our few remaining unit tests need.

In particular, this replacement doesn't offer any means for registering
helper functions to translate struct types with unexported fields. Those
instead just get left set to their zero value in the result. Since our
focus is only on unit test code, we can know ahead of time when that
limitation is relevant and code around it in the calling test rather than
complicating this utility code.