statefile: Use lighter techniques for data copying
This package was our last non-legacy use of the third-party "copystructure"
library, and that library is far more general than most of this work needs.

Most of this is replacing overly-general code with more specific code,
since it's not exactly hard to shallow-copy a map or a slice of strings.

The 2-to-3 upgrade is a little trickier because we were previously copying
the entire data structure before making light modifications to it, since
the V2 and V3 formats are similar enough. However, we can be a little
naughty and use our mainly-test-oriented copy.DeepCopyValue helper instead
of copystructure, since these structures are simple enough for that to
work and are never going to change because they are here only to support
upgrading from long-obsolete file formats.

This removes Terraform's last non-legacy dependency on copystructure, and
with it the last non-legacy indirect dependency on the sibling module
"reflectwalk".