disconnect remote-exec immediately
The remote-exec provisioner was not disconnecting immediately after
provisioning. The remote exec context was tired to the lifetime of the
provisioner, but the provisioner node used to not live long after the
remote execution had completed so it was not noticed. Some more recent
refactoring extends the lifetime of the node, and causes the provisioner
connections to hang around longer when they're not needed.

Ensure we disconnect immediately after the remote execution is complete.