go.mod: go get golang.org/x/crypto@v0.21.0
This update is mostly internal changes that don't affect external behavior.

The only update potentially significant to Terraform is the new exclusion
of a particular root certificate that is used only for some specific names
under the .tr CCTLD, but the certificate database doesn't specify that
constraint in a way that the library can understand, so it's excluded to
avoid making the trust for that certificate broader than intended.