go.mod: go get golang.org/x/net@v0.22.0
This update is largely focused on QUIC implementation changes, which do not
directly affect Terraform.

There is one fix for the HTTP2 implementation to terminate the stream if
a sender recieves an out-of-bounds window update, which makes the
implementation better match the specification. This seems unlikely to
cause any practical problems for Terraform.