go.mod: go get github.com/pkg/browser@5ac0b6a4141c
We use this library as a helper to open a web browser in "terraform login".

The only significant update for Terraform's purposes is that on Windows
it now launches a browser by directly calling the ShellExecuteW function,
instead of running the "start.exe" program that indirectly calls that same
function.