configtesting: Separate package for the SynthBody helper
This "SynthBody" helper doesn't really have anything to do with package
configs except that it's sometimes useful for creating fake instances of
the types in there for testing.

Moving it into its own package means that package backend no longer needs
to depend on package configs, since it doesn't actually have any need for
parsing module configurations.

The old configs.SynthBody function is still available as a forwarding
alias, because most of the existing callers of it are from packages that
already need to import package configs for other reasons, and so this
reduces the need to churn loads of existing tests.