legacy/terraform: Minimize the contents of this legacy package
At one point we were relying on this for a few different things, but now
it's here exclusively to support legacy/helper/schema, the legacy SDK
snapshot that is still being used by a few of the remote state backends.

The SDK only really needs a tiny portion of the functionality from this
package, and by deleting as much of the unused stuff as possible we can
reduce the indirect dependencies of this package and thus, in turn, the
indirect dependencies of the remote state backends.

This also selectively deletes a few parts of the legacy SDK that are
focused on providers and provisioners, because we know that we have this
here only to support the backends that haven't yet been weaned off it,
but we'll save a more adventurous purge of _that_ package for another day.