CODEOWNERS: Comment out entries that GitHub cannot use
GitHub considers any entry that cannot result in an automatic review
request to be invalid.

We want to retain them for human reference, but it's nice to have the file
be considered valid by GitHub so that we can more easily see if we
accidentally make it invalid again in future.