CODEOWNERS: No direct owner for go.sum files
These get updated only as a side-effect of changing a go.mod file, or else
our PR consistency checks would detect that and fail.

The Go toolchain includes in go.sum any module it needed to retrieve to
decide whether or not a module was relevant (whether there's a
specific-package-based indirect dependency) and so if a particular module
sees a go.sum update without an associated go.mod update then that means
that the change cannot possibly affect that module's behavior at runtime.